On branch argPkg Changes to be committed: modified:   autoExport.py modified:   autoExportConfig.ini
Now the config file contains an option to pass command line arguments to
ftPostProc.  With this, the -a/-n/-t command line argument options were
removed from autoExport.py. These options would now be specified in the
config file. This was done to allow support for all the options that
ftPostProc supports, without needing to re-code and pass through from
autoExport.py.